Created own domain / container for eCommerce
