+++ src/Client.php
@@ -12,6 +12,7 @@ use XFApi\Exception\XFApiException;
  * @package XFApi
  *
  * @property XFContainer $xf
+ * @property Container\DBTech\eCommerceContainer $dbtech_ecommerce
  */
 class Client
 {
@@ -21,8 +22,9 @@ class Client
     protected $apiKey;
     protected $apiUserId;
     protected $httpClient;
-
-    protected $_xf;
+	
+	protected $_xf;
+	protected $_dbtech_ecommerce;
 
     /**
      * Client constructor.
@@ -247,18 +249,30 @@ class Client
                 throw new XFApiException('HTTP Error code: ' . $request->getStatusCode());
         }
     }
-
-    /**
-     * @return XFContainer
-     */
-    public function getXf()
-    {
-        if (!$this->_xf) {
-            $this->_xf = new XFContainer($this);
-        }
-
-        return $this->_xf;
-    }
+	
+	/**
+	 * @return XFContainer
+	 */
+	public function getXf()
+	{
+		if (!$this->_xf) {
+			$this->_xf = new XFContainer($this);
+		}
+		
+		return $this->_xf;
+	}
+	
+	/**
+	 * @return Container\DBTech\eCommerceContainer
+	 */
+	public function getDbtechEcommerce()
+	{
+		if (!$this->_dbtech_ecommerce) {
+			$this->_dbtech_ecommerce = new Container\DBTech\eCommerceContainer($this);
+		}
+		
+		return $this->_dbtech_ecommerce;
+	}
 
     /**
      * @param string $name
@@ -267,11 +281,16 @@ class Client
      */
     public function __get($name)
     {
-        $method = 'get' . ucfirst($name);
+        $method = 'get' . $this->camelCase($name);
         if (method_exists($this, $method)) {
             return $this->$method();
         }
 
         throw new XFApiException('Unable to find container ' . $name);
     }
+	
+	protected function camelCase($string, $glue = '_')
+	{
+		return str_replace(' ', '', ucwords(str_replace($glue, ' ', $string)));
+	}
 }
